.history/requirements_20250916020120.txt
segmentation-models-pytorch 
albumentations 
timm
streamlit
torch
torchvision
segmentation-models-pytorch
pillow
opencv-python
numpy
matplotlib